chore: Refactoring stopwatch wrapper.
